ajustes nos testes de juiz e testcase

+++ backend/lets/letscode/model/arquivoSubmissao.py
@@ -2,6 +2,7 @@ import random
 import string
 import os
 
+
 class ArquivoSubmissao():
 
     def __init__(self, codigo):
@@ -10,7 +11,6 @@ class ArquivoSubmissao():
 
     def criarArquivo(self):
         nomeArquivo = self.gerarNomeArquivo()
-        print(nomeArquivo)
         arquivo = open(nomeArquivo,"w+")
 
         return arquivo
@@ -20,25 +20,33 @@ class ArquivoSubmissao():
 
     def apagarArquivo(self):
         #TODO: verificar se o arquivo existe
-        self.arquivo.close()
-        try:
-            print(os.path.realpath(self.arquivo.name))
-            os.remove(os.path.realpath(self.arquivo.name))
-        except OSError:
-            print("maoe")
-        
-        
-        
+        if self.arquivo != None:
+            self.arquivo.close()
+            try:
+                os.remove(os.path.realpath(self.arquivo.name))
+                self.arquivo = None
+            except (OSError):
+                print("disparou")
 
+    def is_arquivo_valido(self):
+        if self.arquivo == None:
+            return False
+
+        return True
+    
     def escreverCodigoNoArquivo(self, codigo):
         codigo = codigo.split("\n")
 
-        
         for linha in codigo:
+            
             # TODO: verificar antes se o arquivo está aberto
             self.arquivo.write(linha+"\n")
 
         self.arquivo.flush()
+
+        print(self.arquivo.read())
+        
+
         return True
 
     # obtido em: https://stackoverflow.com/questions/2257441/random-string-generation-with-upper-case-letters-and-digits-in-python
